+++ tools/cabana/binaryview.h
@@ -6,7 +6,7 @@
 #include <QStyledItemDelegate>
 #include <QTableView>
 
-#include "tools/cabana/dbcmanager.h"
+#include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/streams/abstractstream.h"
 
 class BinaryItemDelegate : public QStyledItemDelegate {
