+++ tools/cabana/dbc/dbc.cc
@@ -1,4 +1,4 @@
-#include "tools/cabana/dbc.h"
+#include "tools/cabana/dbc/dbc.h"
 #include "tools/cabana/util.h"
 
 uint qHash(const MessageId &item) {
