+++ tools/cabana/dbc/dbcmanager.h
@@ -8,8 +8,8 @@
 #include <QSet>
 #include <QDebug>
 
-#include "tools/cabana/dbc.h"
-#include "tools/cabana/dbcfile.h"
+#include "tools/cabana/dbc/dbc.h"
+#include "tools/cabana/dbc/dbcfile.h"
 
 typedef QSet<uint8_t> SourceSet;
 const SourceSet SOURCE_ALL = {};
