+++ tools/cabana/historylog.h
@@ -7,7 +7,7 @@
 #include <QLineEdit>
 #include <QTableView>
 
-#include "tools/cabana/dbcmanager.h"
+#include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/streams/abstractstream.h"
 #include "tools/cabana/util.h"
