+++ tools/cabana/messageswidget.h
@@ -6,7 +6,7 @@
 #include <QSet>
 #include <QTableView>
 
-#include "tools/cabana/dbcmanager.h"
+#include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/streams/abstractstream.h"
 
 class MessageListModel : public QAbstractTableModel {
