+++ tools/cabana/streams/abstractstream.h
@@ -6,7 +6,7 @@
 #include <QColor>
 #include <QHash>
 
-#include "tools/cabana/dbcmanager.h"
+#include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/settings.h"
 #include "tools/cabana/util.h"
 #include "tools/replay/replay.h"
