+++ tools/cabana/tools/findsimilarbits.cc
@@ -8,7 +8,7 @@
 #include <QPushButton>
 #include <QRadioButton>
 
-#include "tools/cabana/dbcmanager.h"
+#include "tools/cabana/dbc/dbcmanager.h"
 #include "tools/cabana/streams/abstractstream.h"
 
 FindSimilarBitsDlg::FindSimilarBitsDlg(QWidget *parent) : QDialog(parent, Qt::WindowFlags() | Qt::Window) {
